Ticket-Generator/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
pydantic>=2.6.4
python-dotenv>=1.0.1
aio-pika>=9.4.0
qrcode>=7.4.2
Pillow>=10.2.0
